Factor out the keyList function duplicated in all test runners
